NP-9149 Add edit button and restructure files
